requirements.txt
streamlit
python-pptx
transformers
torch
sentencepiece
sacremoses
Pillow
pytesseract
pdf2image
ArabicOcr